Add Makefile to setup venv and run the project
Following up PR 22, I had this Makefile lingering which I use to setup
and run Python projects in a venv.

Take note that it also offers an install and update target which can be
used to run an installer and updater if providing the necessary Python
files. There are currently no such files so both targets `make install`
and `make update` will just complain about the missing files.

Link: https://github.com/cakama3a/Polling/pull/22
Signed-off-by: Kai Krakow <kai@kaishome.de>